--- src/main/java/com/drunkenlion/alcoholfriday/domain/admin/customerservice/application/AdminCustomerService.java
@@ -1,4 +1,0 @@
-package com.drunkenlion.alcoholfriday.domain.admin.customerservice.application;
-
-public interface AdminCustomerService {
-}
