+++ src/main/java/com/drunkenlion/alcoholfriday/domain/admin/customerservice/question/application/AdminCustomerService.java
@@ -1,0 +1,4 @@
+package com.drunkenlion.alcoholfriday.domain.admin.customerservice.question.application;
+
+public interface AdminCustomerService {
+}
