+++ src/main/java/com/drunkenlion/alcoholfriday/domain/admin/customerservice/question/application/AdminQuestionService.java
@@ -1,6 +1,6 @@
-package com.drunkenlion.alcoholfriday.domain.admin.customerservice.application;
+package com.drunkenlion.alcoholfriday.domain.admin.customerservice.question.application;
 
-import com.drunkenlion.alcoholfriday.domain.admin.customerservice.dto.response.AdminQuestionResponse;
+import com.drunkenlion.alcoholfriday.domain.admin.customerservice.question.dto.response.AdminQuestionResponse;
 import com.drunkenlion.alcoholfriday.domain.customerservice.dto.request.QuestionModifyRequest;
 import com.drunkenlion.alcoholfriday.domain.member.entity.Member;
 import java.util.List;
