+++ src/test/java/com/drunkenlion/alcoholfriday/domain/admin/customerservice/application/AdminQuestionServiceImplTest.java
@@ -6,7 +6,8 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.Mockito.*;
 
-import com.drunkenlion.alcoholfriday.domain.admin.customerservice.dto.response.AdminQuestionResponse;
+import com.drunkenlion.alcoholfriday.domain.admin.customerservice.question.application.AdminQuestionServiceImpl;
+import com.drunkenlion.alcoholfriday.domain.admin.customerservice.question.dto.response.AdminQuestionResponse;
 import com.drunkenlion.alcoholfriday.domain.customerservice.dao.QuestionRepository;
 import com.drunkenlion.alcoholfriday.domain.customerservice.dto.request.QuestionModifyRequest;
 import com.drunkenlion.alcoholfriday.domain.customerservice.entity.Question;
@@ -20,14 +21,11 @@ import com.drunkenlion.alcoholfriday.global.file.application.FileService;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.ArgumentMatchers;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
